+++ src/stories/atoms/Logo/index.ts
@@ -1,0 +1,3 @@
+export { Logo } from './Logo';
+export type { LogoProps, LogoSize } from './Logo';
+export { logoPaths, logoSvg, logoViewBox } from './logoAssets';